refactor `additional` subpackage (move `FreeCADTemplates` up)
