+++ freecad_stubs/FreeCADTemplates/templates.py
