+++ freecad_stubs/FreeCADTemplates/type_consts.py
