+++ lib/freecad_stub_gen/FreeCADTemplates/__init__.py
