+++ lib/freecad_stub_gen/FreeCADTemplates/type_consts.py
